scin_legacy/model/requirements.txt
# PyTorch 및 Torchvision
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# 머신러닝 유틸리티
scikit-learn>=1.3.0
tqdm>=4.65.0
efficientnet-pytorch>=0.7.1

# 시각화
matplotlib>=3.7.0
seaborn>=0.12.0

# TensorBoard (학습 모니터링)
tensorboard>=2.13.0

# 기타 유틸리티
pyyaml>=6.0
